FIX: Allow compiling for Mac OS X (Zhaop)

+++ gui/simwin.cc
@@ -498,7 +498,7 @@ void rdwr_all_win(loadsave_t *file)
 				switch(id) {
 
 					// end of dialogues
-					case magic_none: return;
+					case (uint32)magic_none: return;
 
 					// actual dialogues to restore
 					case magic_convoi_info:    w = new convoi_info_t(); break;
